requirements.txt
# LeetCode to GitHub Sync - Dependencies

# HTTP requests
requests==2.31.0

# GitHub API
PyGithub==2.1.1

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Date/time utilities
python-dateutil==2.8.2

# Optional: Better logging
colorlog==6.8.0
